requirements.txt
aiochclient==2.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiokafka==0.12.0
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.7.14
click==8.2.1
clickhouse-driver==0.2.9
fastapi==0.115.14
frozenlist==1.7.0
h11==0.16.0
idna==3.10
kafka-python==2.2.15
multidict==6.6.3
packaging==25.0
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
sentry-sdk==2.34.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.46.2
typing-extensions==4.14.1
typing-inspection==0.4.1
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
